Update ER32_CrateSpawner.sqf
Arguments were missing, breaking the whole "crateSpawner.sqf" script. Fixed it by adding the arguments back.

Used "if () do {}" ... that of course does not work. Changed to "if () then {}" which will work. :)

Ressources were only added correctly for the first crate spawned. After spawning a second crate, all crates would be empty.
The array for the ressources were not properly copied. Fixed it.